move themer dropdown controls to index controller
